+++ cloudfeaster/samples/xe_exchange_rates.py
@@ -44,8 +44,9 @@ class XEExchangeRatesSpider(spider.Spider):
 
 
 if __name__ == '__main__':
+    crawl_debugger = spider.CrawlDebugger()
     crawl_args = spider.CLICrawlArgs(XEExchangeRatesSpider)
-    crawler = spider.SpiderCrawler(XEExchangeRatesSpider)
+    crawler = spider.SpiderCrawler(XEExchangeRatesSpider, crawl_debugger.debug)
     crawl_result = crawler.crawl(*crawl_args)
     print json.dumps(crawl_result)
     sys.exit(1 if crawl_result.status_code else 0)